Create useAuthCheck custom hook
Previously when refreshing or navigating in a new session, `PrivateRoute` immediately redirected you to the home screen. That's because on load the auth state is false and the logic built into that component ran before the `checkAuth()` effect in `App`.

This was annoying in development and ultimately not the behavior that a user would want. So I made the following changes:

First, I made a destinction in the App state between "authenticated", "unauthenticated", and "not authenticated" by defaulting `isAuthenticated` to null. I coupled that with a specific DEAUTHENTICATE action.

Next I created the custom hook that runs on mount. It might be safe to remove the dependency array, but it is working as is. I added this `useAuthCheck` hook to `App` and `PrivateRoute`.

Finally, I made the distinction in `PrivateRoute` between the auth states. The component waits for a response from /refresh before making a decision.

The only potential problem I see with this solution is that it is dependent on latency. If that becomes an issue, then this solution will need reworked.